Refactor TopBar
