+++ frontend/src/components/TopBar/index.tsx
@@ -1,0 +1,3 @@
+import TopBar from "./TopBar";
+
+export default TopBar;
